SDA/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bert-score==0.3.13
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
faiss-gpu==1.7.2
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.2.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
openai==1.59.7
packaging==24.2
pandas==2.2.3
pillow==10.2.0
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge==1.0.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.48.0
triton==3.1.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
zipp==3.21.0
